# storage: do not report WriteConflict or PessimisticLockNotFound if txn is committed