@sticnarf 2021-07-06
> It's better to do some benchmarks under high-conflict scenario.

Yes. I will do some benchmark on high-conflict optimistic workloads.

PessimisticLockNotFound occurs rarely even in high-conflict workloads, so I think we needn't worry about that.