@sticnarf 2021-07-08
> How about the scheduler-prewrite panel? I want to confirm the cursor operation.

Like this 
![图片](https://user-images.githubusercontent.com/17217495/124921727-47568480-e02b-11eb-88b4-6f0deb1d08d6.png)

I spent whole day finding why there was no "next" cursor operations until I found statistics are not updated on errors  T_T

In this PR I only fix it for prewrite with a small change. I'd like to fix for all commands with some better ways.