@youjiali1995 2021-07-16
We should cherry-pick it and https://github.com/tikv/tikv/pull/10522 to release-5.0/5.1.